server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9
torch>=2.5.0
torchvision>=0.20.0
segmentation-models-pytorch>=0.3.0
Pillow>=11.0.0
numpy>=1.26.0
coremltools>=8.0
